Refactor ES tests to remove ES version abstraction
